Dataset: Split module for handling dataset IO to filesystem
Also refactor imports and tests.
